Adds custom error region function to pick up Irony's end column reporting.
If Irony can report the end column for a diagnostic message, then it's
much simpler, and faster, to use that to calculate the error region than
what Flycheck itself can do - especially if mode is `sexps'. It will
also be more accurate.

This adds a custom `flycheck-irony-error' struct that inherits from
`flycheck-error' and adds a `column-end' member.

`flycheck-irony--build-error' has been changed to build a
`flycheck-irony-error' instead of a `flycheck-error' and set `column-end'
based on Irony's diagnostic output.

A new error region function has been added; `flycheck-irony--error-region'.
That function is added as an advice around the standard
`flycheck-error-region-for-mode'. If the checker is "irony" and
`column-end' has as value, `flycheck-irony--error-region' will calculate
the error region, disregarding mode. If not, it will call the standard
`flycheck-error-region-for-mode'.